Refactored to use redux instead of context
